requirements.txt
httpx
invoke
pyjwt
pyyaml
